requirements.txt
python-interface==1.5.0
numpy==1.15.0
scipy==1.1.0
tensorflow-gpu==1.9.0